Add files via upload
Previously when I submitted the assessment in 1 day I couldn't implement the Material UI properly in the carousel section as it was my first time using it and I forgot to implement the Hamburger menu in the responsive design. I've made those changes and made it look much better and appropriate according to the Figma image provided.